refactor: implementing new verify token page
